Chore: Separated functional tests in own module
